requirements.txt
# 基础依赖
smolagents==1.21.3
dashscope>=1.10.0  # 通义千问模型依赖
langchain==0.3.27
chromadb==1.1.0
transformers==4.56.1
pydantic==2.11.9

# 文档处理
tiktoken==0.11.0
PyPDF2==3.0.0
python-docx==1.2.0

# 网络请求
requests==2.32.5
ddgs==9.6.0

# 向量存储
faiss-cpu==1.8.0

# 其他
tqdm==4.67.1